Implemented and tested.  The issue was in the exception handler call back.  Tested against upload, download, backup, all functions.  Fixes #113.
